frontend/requirements.txt
streamlit
requests
python-dotenv
fpdf2
langfuse==2.60.0
# only works with python 3.12 